src/requirements.txt
click==7.1.2
cycler==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.2
nltk==3.5
numpy==1.19.0
pyparsing==2.4.7
PySimpleGUI==4.20.0
python-dateutil==2.8.1
regex==2020.6.8
six==1.15.0
tqdm==4.46.1
